refactor(storage): move embedding helpers out of lib
