+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 580
+- **Stable** (no move scoped): 579
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (lib-root + storage-root): 29
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 697
-- **Realized polylogue/**/*.py**: 697 files declared
+- **Total declared**: 696
+- **Realized polylogue/**/*.py**: 696 files declared
 <!-- END GENERATED: topology-status -->
