+++ polylogue/storage/embeddings/__init__.py
@@ -1,0 +1,25 @@
+"""Embedding storage, materialization, and readiness helpers."""
+
+from polylogue.storage.embeddings.materialization import (
+    EmbedConversationOutcome,
+    EmbedSingleStatus,
+    PendingConversation,
+    embed_conversation_sync,
+    iter_pending_conversations,
+)
+from polylogue.storage.embeddings.status_payload import (
+    EmbeddingStatusPayload,
+    RetrievalBandPayload,
+    embedding_status_payload,
+)
+
+__all__ = [
+    "EmbedConversationOutcome",
+    "EmbedSingleStatus",
+    "EmbeddingStatusPayload",
+    "PendingConversation",
+    "RetrievalBandPayload",
+    "embed_conversation_sync",
+    "embedding_status_payload",
+    "iter_pending_conversations",
+]
