+++ polylogue/storage/embeddings/status_payload.py
